==> 1a-AcquiringPostID.ipynb <==
# 1-Acquiring Post ID
## Author: Cassie Buhler
### Last modified: November 29th 2020 
###Email: cb3452@drexel.edu

About: In this code I use pushshift to obtain Reddit post IDs for the subreddit AITA. The post IDs are saved as a .json file. 



import requests
import json
import time 
import datetime

from google.colab import drive
drive.mount('/content/gdrive')


def getIDs(start, end):
    url = 'https://api.pushshift.io/reddit/submission/search/?subreddit=amitheasshole&after='+ str(start) +"&before"+str(end)+"&size=1000"
    request = requests.get(url).text
    posts = json.loads(request)
    return posts['data']


start = 1420070460 # epoch for January 1 2015
end = 1604188860 # epoch for November 1 2020
ids = []

firstDay = start
while int(firstDay) < end: 
    posts = getIDs(firstDay,end) #calling function until we get all the posts in the time window 
    for post in posts:
        ids.append(post['id'])
    firstDay = post['created_utc']
    print(len(ids))
    print(datetime.datetime.fromtimestamp(int(firstDay)).strftime('%Y-%m-%d %H:%M:%S'))
    time.sleep(0.5)



path = "/content/gdrive/My Drive/DSCI511/FinalProject/data/"

with open(path + 'postIDs.json','w') as f:
  json.dump(ids,f)
