==> Part2-CreateDataset.ipynb <==
# Part2-CreateDataset
### Author: Cassie Buhler
### Last modified: November 29th 2020 
### Email: cb3452@drexel.edu

##### About: Given postID, I use PRAW to obtain Reddit post features. These features are saved to as a dataset.


!pip install praw

import praw
import re
import json
import pandas as pd
from collections import Counter
import datetime
import time

from google.colab import drive
drive.mount('/content/gdrive')

def getPRAW():
  path = "/content/gdrive/My Drive/DSCI511/FinalProject/data/"
  with open(path + 'myKeys.json') as f: #REPLACE WITH YOUR PRAW KEYS HERE 
      keys = json.load(f)

  access = praw.Reddit(client_id = keys['client_id'],
                      client_secret = keys['client_secret'],
                      user_agent = keys['user_agent'],
                      username = keys['username'] ,
                      password =  keys['password'])
  return access

def makeDataFrame(title,content,numUpvote,upvoteRatio,numComments,links,ids, dates,judgements,numAwards):
  dataframe = pd.DataFrame()
  dataframe["Judgement"] = judgements
  dataframe['Title'] = title
  dataframe['Content'] = content
  dataframe['Date'] = dates
  dataframe['Link'] = links
  dataframe['ID'] = ids
  dataframe['numComments'] = numComments
  dataframe['numUpvotes'] = numUpvote
  dataframe['upvoteRatio'] = upvoteRatio
  dataframe['numAwards'] =numAwards
  return dataframe

def getPostData(ids):
  content = []
  title =[]
  numComments = []
  upvoteRatio = []
  numUpvote = []
  dates = []
  judgements = []
  awards =[]
  count = 0
  numAwards= 0
  links = []
  postId = []
  emptyList = ["", "[deleted]", "[removed]"]
  verdictList = ["asshole","not the asshole","everyone sucks","no assholes here"]
  for id in ids:
    access = getPRAW()
    post = access.submission(id = id)
    count +=1
    print(count)
    if post.link_flair_text in verdictList:
      if post.selftext not in emptyList:
        postId.append(id)
        dates.append(datetime.datetime.fromtimestamp(int(post.created_utc)).strftime('%Y-%m-%d %H:%M:%S'))
        links.append(post.url)
        judgements.append(post.link_flair_text)
        title.append(post.title)
        numComments.append(post.num_comments)
        numUpvote.append(post.score)
        content.append(post.selftext)
        upvoteRatio.append(post.upvote_ratio)
        for award in post.all_awardings:
          numAwards += award['count']
        awards.append(numAwards)
    if count%1000 == 0:
      dataframe = makeDataFrame(title,content,numUpvote,upvoteRatio,numComments,links,postId,dates,judgements,awards)
      dataframe.to_csv("/content/gdrive/My Drive/DSCI511/FinalProject/data/temporaryData_" + str(count) + ".csv")
  dataframe = makeDataFrame(title,content,numUpvote,upvoteRatio,numComments,links,postId,dates,judgements,awards)
  return dataframe 

path = "/content/gdrive/My Drive/DSCI511/FinalProject/data/"
with open(path + 'postIDs.json') as f:
  ids = json.load(f)
reducedIDs = ids[1::10] #It takes about 70 seconds per 100 posts so I need to reduce the datasize since this would take me 6 days to run this on the entire dataset 

data = getPostData(reducedIDs)
data.to_csv(path+ "Data.csv")
