==> Part3-CleanData.ipynb <==
# Part3-CleanData
## Author: Cassie Buhler
### Last modified: November 30th 2020 
###Email: cb3452@drexel.edu


import pandas as pd

from google.colab import drive
drive.mount('/content/gdrive')


path1 = "/content/gdrive/My Drive/DSCI511/FinalProject/data/"
path2 = "/content/gdrive/My Drive/DSCI511/FinalProject/data-big/"


olderData = pd.read_csv(path1 + 'temporaryData_1000.csv')
olderData.drop(columns = ['Unnamed: 0'],inplace = True)
olderData['Judgement'].replace({'not the asshole': 'NTA', 'asshole': 'YTA', 'everyone sucks': 'ESH', 'no assholes here': 'NAH'},inplace = True)


YTA = pd.read_csv(path2 + 'YTA.csv')
NTA = pd.read_csv(path2 + 'NTA.csv')
ESH = pd.read_csv(path2 + 'ESH.csv')
NAH = pd.read_csv(path2 + 'NAH.csv')

def cleanData(data,label):
  data.drop(columns = ['Unnamed: 0','YTAcount','NAHcount','NTAcount','ESHcount'],inplace = True)
  awardcols = data.iloc[:,7:]
  numAwards = awardcols.sum(axis = 1)
  data.drop(columns = awardcols.columns, inplace = True)
  data['numAwards'] = numAwards.astype(int)
  data['Judgement'] = [label]*len(data)
  data.reset_index()
  return data

YTA = cleanData(YTA,'YTA')
NTA = cleanData(NTA,'NTA')
ESH = cleanData(ESH,'ESH')
NAH = cleanData(NAH,'NAH')

temp1 = pd.concat([YTA, NTA])
temp2 = pd.concat([ESH, NAH])
temp = pd.concat([temp1, temp2])
mergedData = olderData.merge(temp, how = 'outer')
mergedData.drop(columns = 'ID', inplace = True)

mergedData.to_csv(path1+'Data_cleaned.csv')

mergedData

path1 = "/content/gdrive/My Drive/DSCI511/FinalProject/data/"
import pandas as pd
pd.read_csv(path1+'Data_cleaned.csv')
